[TMP] circleci: Upload coverage reports to codecov.io.
TMP: Remove Travis config, simultaneously, so codecov gives a comparison.
TMP: try pip package after all -- with --user

The Travis config does this by a `pip install codecov` in setup and
running the resulting `codecov` command at the end.  That risks
confusion because it makes the library look like something we might
be using *during* the tests to collect coverage information; and
the README for the codecov Python client says the bash client is
preferred.  So, use the bash client instead, with the recommended
command for downloading and running it.

Also, because this step is inherently about talking to a
third-party service whose reliability we can't control, and because
it's purely a reporting step that happens after our actual test
suite has succeeded, don't fail if it fails; just print a message.
(I'd prefer a somewhat louder message, like turning that step
yellow or something in the Circle web UI, but Circle doesn't offer
an option like that.  If the coverage reports start consistently
failing for some reason, then if we're actually reading them we'll
notice anyway.)

[greg: Edited step name and error message; wrote this commit message.]