migrations: Hack to drop NOT NULL on fields we're going to ignore.
When we remove these fields from the model definition, calls to the
model's .objects.create() method will generate INSERT queries that
leave them out.  That will cause the database to reject the INSERT
with an error if the columns are marked NOT NULL.

So, drop that constraint first.  Then in the next few commits we can
drop the fields from the model definition, along with the code that
tries to use them.  Then when that's deployed, it'll be safe to run
the migrations that actually drop these columns from the database,
which finally will allow us to actually deploy the new code.

This would be a migration, except I didn't want to confuse the
migrations system by having this migration appearing in one deploy and
vanishing in the next.  So instead, it's a management command that
must be run before deploying the next few commits; and so I could test
this whole branch and get a passing result, it's slipped into test
setup methods.  We seem to have a nasty habit of not calling `super`
in our numerous `setUp` methods; but fortunately we have very few
`setUpClass` methods, so sticking it in that one on ZulipTestCase
covers the bulk of our tests.