wip typing: cut "empty message" condition
TODO oh but what if there's a recipient entered; you typed something;
  then abandoned; then start by deleting?  Not unreasonable to
  suppress a typing notif... though otoh if you're not right about to
  type something else, why didn't you simply close the box; also if it
  takes you more than one keystroke to delete, this doesn't suppress
  anyway.

This conditional already has a very narrow effect: we only call this
function in the first place when (a) the message content input has
just been edited, and (b) the set of PM recipients (empty, if
composing a stream message) is different from the last time the
content input was edited.

In particular, when the user is in the middle of editing the set of
recipients, this isn't called.  And when they're in the middle of
editing the message, it isn't called.  It's called, at most, on the
first edit the user makes to the content after a change to the
recipients.

The effect of the conditional is that if that one first edit makes the